point2_1 Created styles for Chip and apply these styles
