+++ app/src/main/java/geekbarains/material/view/ui/fragmetns/PictureOfTheDayFragment.kt
@@ -92,7 +92,7 @@ class PictureOfTheDayFragment : Fragment() {
             R.id.app_bar_fav -> {
                 toast("Favourite")
             }
-            R.id.app_bar_settings -> activity?.supportFragmentManager?.beginTransaction()?.add(R.id.container, ChipsFragment())?.addToBackStack(null)?.commit()
+            R.id.app_bar_settings -> activity?.supportFragmentManager?.beginTransaction()?.add(R.id.container, SettingFragment())?.addToBackStack(null)?.commit()
             android.R.id.home -> {
                 activity?.let {
                     BottomNavigationDrawerFragment().show(it.supportFragmentManager, "tag")
