+++ app/src/main/res/layout/fragment_setting.xml
@@ -1,12 +1,48 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
     android:orientation="vertical">
 
-    <LinearLayout
+    <TextView
+        android:id="@+id/tv_choose_theme"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_margin="@dimen/default_offset_full"
+        android:text="@string/tv_text_choose_theme"
+        android:textSize="18sp"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintLeft_toLeftOf="parent"/>
+
+    <com.google.android.material.chip.ChipGroup
+        android:id="@+id/cg_choose_theme"
+        android:layout_margin="@dimen/default_offset_full"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        app:singleSelection="true"
+        app:layout_constraintTop_toBottomOf="@id/tv_choose_theme"
+        app:layout_constraintStart_toStartOf="parent">
+
+        <com.google.android.material.chip.Chip
+            style="@style/ChipSelectTheme"
+            android:layout_height="wrap_content"
+            android:text="@string/chip_text_theme_day" />
+
+        <com.google.android.material.chip.Chip
+            style="@style/ChipSelectTheme"
+            android:layout_height="wrap_content"
+            android:text="@string/chip_text_theme_night" />
+
+        <com.google.android.material.chip.Chip
+            style="@style/ChipSelectTheme"
+            android:layout_height="wrap_content"
+            android:text="@string/chip_text_theme_toxic" />
+    </com.google.android.material.chip.ChipGroup>
+
+    <!--LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="horizontal">
@@ -94,37 +130,7 @@
             android:text="@string/chip_text_two" />
     </com.google.android.material.chip.ChipGroup>
 
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_margin="8dp"
-        android:text="@string/textview_text_choose_one"
-        android:textSize="18sp" />
 
-    <com.google.android.material.chip.ChipGroup
-        android:id="@+id/chipGroup"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:singleSelection="true">
-
-        <com.google.android.material.chip.Chip
-            style="@style/Widget.MaterialComponents.Chip.Choice"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/chip_text_choice_a" />
-
-        <com.google.android.material.chip.Chip
-            style="@style/Widget.MaterialComponents.Chip.Choice"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/chip_text_choice_b" />
-
-        <com.google.android.material.chip.Chip
-            style="@style/Widget.MaterialComponents.Chip.Choice"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/chip_text_choice_c" />
-    </com.google.android.material.chip.ChipGroup>
 
     <HorizontalScrollView
         android:layout_width="match_parent"
@@ -200,5 +206,5 @@
         android:layout_height="wrap_content"
         android:layout_gravity="center_horizontal"
         android:layout_marginTop="24dp"
-        android:text="@string/chip_text_close_icon_listener" />
-</LinearLayout>
+        android:text="@string/chip_text_close_icon_listener" /-->
+</androidx.constraintlayout.widget.ConstraintLayout>
