+++ app/src/main/res/values/colors.xml
@@ -7,8 +7,8 @@
     <color name="color_black">#FF000000</color>
     <color name="bottom_menu_dg">#FFF6F6FA</color>
     <color name="white">#FFFFFF</color>
-    <color name="info_bg">#663b4351</color>
+    <color name="info_bg">#663B4351</color>
 
 
-    <color name="ic_launcher_background">#FFFFFF</color>
+    <color name="fragment_background">#FFFFFF</color>
 </resources>
